+++ Source/Entities/DoorEntity.cpp
@@ -34,7 +34,7 @@ namespace Freeking
 			Close();
 		}
 
-		_currentDistance += ((_speed * dt) * (_open ? 1.0f : -1.0f));
+		_currentDistance += ((_speed * (float)dt) * (_open ? 1.0f : -1.0f));
 		_currentDistance = Math::Clamp(_currentDistance, 0.0f, _distance);
 
 		SetPosition(_initialPosition.MulAdd(_currentDistance, Vector3f::Up));
